+++ src/v1/BasicTypes/Objects/getProperty.ts
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2020-present Ganbaro Digital Ltd
+// Copyright (c) 2022-present Ganbaro Digital Ltd
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -32,25 +32,29 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./AttributeFilterMap";
-export * from "./AttributeKeys";
-export * from "./AttributeTransformerMap";
-export * from "./DeepImmutable";
-export * from "./Definitely";
-export * from "./EquivalentKeys";
-export * from "./EquivalentOptionalKeys";
-export * from "./EquivalentOptionalPart";
-export * from "./EquivalentPart";
-export * from "./IdenticallyNamedKeys";
-export * from "./IdenticallyNamedPart";
-export * from "./IfEquals";
-export * from "./Immutable";
-export * from "./Maybe";
-export * from "./NonNullable";
-export * from "./Nullable";
-export * from "./OptionalKeys";
-export * from "./OptionalPart";
-export * from "./RequireAllAttributesMap";
-export * from "./ValueOf";
-export * from "./WritableKeys";
-export * from "./WritablePart";
+/**
+ * `getProperty()` retrieve the property called `propName` from the
+ * given `target` object.
+ *
+ * This is a workaround added because Typescript does not support
+ * symbols as object property names out of the box. You can call this
+ * function to avoid adding linter comments everywhere in your own code.
+ *
+ * @param target -
+ * The object that we are retrieving from.
+ * @param propName -
+ * The property that we are looking for.
+ *
+ * @typeParam T -
+ * The type of the returned value.
+ *
+ * @returns
+ * - `target[propName]` if it exists
+ * - `undefined` otherwise
+ */
+export function getProperty<T = unknown>(
+     target: object,
+     propName: string|number|symbol
+): T {
+    return (target as Record<typeof propName, any>)[propName as any] ?? undefined;
+}