+++ src/v1/SupportingTypes/DispatchMap/searchDispatchMap.ts
@@ -35,6 +35,8 @@
 
 import { AnyFunction } from "../../Archetypes";
 import { DispatchMap } from "./DispatchMap";
+import { AnyDispatchMapKey } from "./AnyDispatchMapKey";
+import { getProperty, hasProperty } from "../../BasicTypes";
 
 /**
  * `searchDispatchMap()` will iterate through `keysToTry`
@@ -61,16 +63,19 @@ import { DispatchMap } from "./DispatchMap";
  *
  * @public
  */
- export function searchDispatchMap<F extends AnyFunction>(
-    table: DispatchMap<any, F>,
-    keysToTry: string[],
+ export function searchDispatchMap<F extends AnyFunction, K extends AnyDispatchMapKey>(
+    table: DispatchMap<K, F>,
+    keysToTry: AnyDispatchMapKey[],
     fallback: F,
 ): F {
     // do we have any of the requested keys?
+    //
+    // Typescript does not support using symbols as object property
+    // names here. We use a workaround to avoid littering our code
+    // with casts to `any`.
     for (const keyToTry of keysToTry) {
-        if (table[keyToTry]) {
-            // we have a winner!
-            return table[keyToTry];
+        if (hasProperty(table, keyToTry)) {
+            return getProperty(table, keyToTry);
         }
     }